refactor: 디렉토리 구조 변경 및 페이지 컴포넌트화
