+++ src/components/Layout/AppLayout.jsx
@@ -2,6 +2,10 @@ import React from 'react';
 import { Box, Flex, IconButton } from '@chakra-ui/react';
 import { FaHome, FaSearch, FaTshirt, FaUser } from 'react-icons/fa'; // 아이콘
 import { BrowserRouter as Router, Route, Routes, Link } from 'react-router-dom';
+import Home from '../../pages/Home';
+import ClothingSearch from '../../pages/ClothingSearch';
+import PartySearch from '../../pages/PartySearch';
+import MyClothes from '../../pages/MyClothes';
 
 function AppLayout() {
   return (
@@ -65,27 +69,6 @@ function AppLayout() {
       </Router>
     </>
   );
-}
-
-function Home() {
-  return <div>홈 화면</div>;
-}
-
-function PartySearch() {
-  return <div>파티 찾기</div>;
-}
-
-function ClothingSearch() {
-  //   return <div>옷 찾기</div>;
-  return (
-    <div>
-      <IconButton icon={<FaUser />} />
-    </div>
-  );
-}
-
-function MyClothes() {
-  return <div>내 옷</div>;
 }
 
 export default AppLayout;
