+++ src/main.jsx
@@ -1,7 +1,7 @@
 import { createRoot } from 'react-dom/client';
-import App from './App.jsx';
+import AppExample from './examples/AppExample.jsx';
 import { ChakraProvider } from '@chakra-ui/react'; // ChakraProvider 추가
-import AppLayout from './AppLayout.jsx';
+import AppLayout from './components/Layout/AppLayout.jsx';
 
 createRoot(document.getElementById('root')).render(
   <ChakraProvider>
