+++ src/pages/ClothingSearch.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function ClothingSearch() {
+  return (
+    <>
+      <div>ClothingSearch</div>
+    </>
+  );
+}
+
+export default ClothingSearch;
