+++ src/pages/Home.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import { Box, Flex, Text, Button } from '@chakra-ui/react';
+
+function Home() {
+  return (
+    <>
+      <div>home</div>
+    </>
+  );
+}
+
+export default Home;
