+++ src/pages/MyClothes.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function MyClothes() {
+  return (
+    <>
+      <div>MyClothes</div>
+    </>
+  );
+}
+
+export default MyClothes;
