+++ src/pages/PartySearch.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function PartySearch() {
+  return (
+    <>
+      <div>PartySearch</div>
+    </>
+  );
+}
+
+export default PartySearch;
