+++ src/utils/firebaseConfig.js
